+++ src/main/java/io/github/astrapi69/entity/id/generator/IntegerIdGenerator.java
@@ -26,38 +26,56 @@ package io.github.astrapi69.entity.id.generator;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
+import lombok.NonNull;
+import io.github.astrapi69.data.identifiable.IdGenerator;
+
 /**
- * The class {@link SystemTimeIdGenerator} is an base implementation of {@link IdGenerator}
- * interface
- *
+ * The class {@link IntegerIdGenerator} is an implementation of {@link IdGenerator} interface with
+ * id type of {@link Integer} object
+ * 
  * @deprecated use instead the same name class from the module id-generate
  */
-public class SystemTimeIdGenerator implements IdGenerator
+public class IntegerIdGenerator implements IdGenerator<Integer>
 {
 
 	/**
-	 * The instance.
+	 * The default instance with an initial value of 0
 	 */
-	private static final SystemTimeIdGenerator instance = new SystemTimeIdGenerator();
-	/**
-	 * The atomic id.
-	 */
-	private final AtomicInteger atomicId;
+	private static final IntegerIdGenerator instance = new IntegerIdGenerator(0);
 
 	/**
-	 * Instantiates a new system time id generator.
+	 * The atomic id counter
 	 */
-	private SystemTimeIdGenerator()
+	private final AtomicInteger atomicIdCounter;
+
+	/**
+	 * Instantiates a new {@link IntegerIdGenerator}
+	 *
+	 * @param initialValue
+	 *            the initial value for the generator
+	 */
+	private IntegerIdGenerator(final @NonNull Integer initialValue)
 	{
-		atomicId = new AtomicInteger((int)System.currentTimeMillis());
+		atomicIdCounter = new AtomicInteger(initialValue);
 	}
 
 	/**
-	 * Gets the single instance of SystemTimeIdGenerator.
+	 * Factory method for create a new custom {@link IntegerIdGenerator} with an initial value
 	 *
-	 * @return single instance of SystemTimeIdGenerator
+	 * @param initialValue
+	 *            the initial value for the generator
 	 */
-	public static SystemTimeIdGenerator getInstance()
+	public static IntegerIdGenerator of(final @NonNull Integer initialValue)
+	{
+		return new IntegerIdGenerator(initialValue);
+	}
+
+	/**
+	 * Gets the single instance of {@link IntegerIdGenerator} object
+	 *
+	 * @return single instance of {@link IntegerIdGenerator} object
+	 */
+	public static IntegerIdGenerator getInstance()
 	{
 		return instance;
 	}
@@ -66,13 +84,8 @@ public class SystemTimeIdGenerator implements IdGenerator
 	 * {@inheritDoc}
 	 */
 	@Override
-	public int getNextId()
+	public Integer getNextId()
 	{
-		int nextId = atomicId.getAndIncrement();
-		if (nextId < 0)
-		{
-			nextId *= -1;
-		}
-		return nextId;
+		return atomicIdCounter.getAndIncrement();
 	}
 }