+++ CHANGELOG.md
@@ -4,10 +4,23 @@
 Version 3.2-SNAPSHOT
 -------------
 
+ADDED:
 
+- new lombok annotations for @EqualsAndHashCode and @ToString to all subclasses that call super
+
+CHANGED:
+
+- tagged all classes IdGenerator as deprecated, classes are moved to own module id-generate
 
 Version 3.1
 -------------
+
+ADDED:
+
+- new class AtomicBigInteger for generator class BigIntegerIdGenerator
+- new generator class BigIntegerIdGenerator for id generation of type BigInteger
+- new generator class IntegerIdGenerator for id generation of type Integer
+- new generator class LongIdGenerator for id generation of type Long
 
 CHANGED: